refactor: src/fiber 내부 디렉토리 분기
Desc
- 함수를 기능 별로 구분하기 위해 디렉토리 분기 수행 (각각 fiber/core, fiber/state로 분기)
