+++ src/fiber/core/createFiberNode.ts
@@ -1,5 +1,5 @@
 import { VirtualNode } from "@/jsx/type.jsx";
-import { FiberNode } from "./type.fiber";
+import { FiberNode } from "../type.fiber";
 
 export function createFiberNode(virtualNode: VirtualNode): FiberNode {
   return {
